requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
rich>=13.7.0
textual>=0.58.1
